fix: use image hash instead of image ID in WLID map
Previously, the ImageID to WLIDs map used values that were sort of like
Image IDs, but not really. This is problematic, since the map is
supposed to be used as a way to look up entities that have Image Hashes
as their keys.

When receiving such an entity, for example, an SBOM, our code cannot
meaningfully transform the Image Hash of an SBOM, to an unambiguous,
correct Image ID so we can look it up in the Image-WLID map.

Therefore, this map should use Image Hashes instead.

This commit refactors operations on the map to insert by Image Hash
instead of Image IDs.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>