fix: Use image hash in WLID map (#94)
* feat(watcher): implement a thread-safe ImageID to WLIDs map

This data structure is designed to replace the map-and-mutex that are
side-by-side in the WatchHandler so clients can perform thread-safe
operations on it with more confidence and ease.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* refactor(watcher): use in IWMap in WatchHandler’s unexported functions

As the first step in wiring the refactoring, use the IWMap in
WatchHandler’s private functions that previously handled thread-safe
operations on the raw ImageIDs to WLIDs map

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* refactor: use IWMap contains check instead of a function

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* refactor(watcher): fix static analysis warnings in IWMap

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* refactor(watcher): remove unused unexported function

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* refactor(watcher): move private cleanup func to IWMap method calls

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* refactor: rename mentions of image ID to image hashes

The data structure maps image hashes to WLIDs. Image hashes are not the
same as image IDs. Image hashes are plain hash digest values. Rename
mentions to avoid confusion.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* tests(watcher): use explicit image hash values

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

* fix: use image hash instead of image ID in WLID map

Previously, the ImageID to WLIDs map used values that were sort of like
Image IDs, but not really. This is problematic, since the map is
supposed to be used as a way to look up entities that have Image Hashes
as their keys.

When receiving such an entity, for example, an SBOM, our code cannot
meaningfully transform the Image Hash of an SBOM, to an unambiguous,
correct Image ID so we can look it up in the Image-WLID map.

Therefore, this map should use Image Hashes instead.

This commit refactors operations on the map to insert by Image Hash
instead of Image IDs.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>

---------

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>